fix: middleware edge
